+++ src/main/java/cz/thepatik/welcomeplugin/commands/SubCommandPlayer.java
@@ -1,16 +1,18 @@
 package cz.thepatik.welcomeplugin.commands;
 
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
 public abstract class SubCommandPlayer {
 
     public abstract String getName();
-    public abstract String getPermissions();
+    public abstract @NotNull String getPermissions();
     public abstract String getDescription();
     public abstract String getSyntax();
     public abstract boolean hasArguments();
-    public abstract List<String> tabComplete(String[] args);
+    public abstract List<String> tabComplete(Player p, String[] args);
+
     public abstract void perform(Player player, String args[]);
 }