+++ src/main/java/cz/thepatik/welcomeplugin/commands/subcommands/player/PlayedTimeCommand.java
@@ -5,6 +5,7 @@ import cz.thepatik.welcomeplugin.commands.SubCommandPlayer;
 import me.clip.placeholderapi.PlaceholderAPI;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Collections;
 import java.util.List;
@@ -22,7 +23,7 @@ public class PlayedTimeCommand extends SubCommandPlayer {
     }
 
     @Override
-    public String getPermissions() {
+    public @NotNull String getPermissions() {
         return "welcomeplugin.playedtime";
     }
 
@@ -41,7 +42,7 @@ public class PlayedTimeCommand extends SubCommandPlayer {
         return true;
     }
     @Override
-    public List<String> tabComplete(String[] args){
+    public List<String> tabComplete(Player p, String[] args){
         return Collections.emptyList();
     }
     @Override