+++ src/main/java/cz/thepatik/welcomeplugin/commands/subcommands/player/SentMessagesCommand.java
@@ -5,6 +5,7 @@ import cz.thepatik.welcomeplugin.commands.SubCommandPlayer;
 import me.clip.placeholderapi.PlaceholderAPI;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Collections;
 import java.util.List;
@@ -20,7 +21,7 @@ public class SentMessagesCommand extends SubCommandPlayer {
     }
 
     @Override
-    public String getPermissions() {
+    public @NotNull String getPermissions() {
         return "welcomeplugin.sentmessages";
     }
 
@@ -40,7 +41,7 @@ public class SentMessagesCommand extends SubCommandPlayer {
     }
 
     @Override
-    public List<String> tabComplete(String[] args) {
+    public List<String> tabComplete(Player p, String[] args) {
         return Collections.emptyList();
     }